patent_engine/requirements.txt
openai
sqlalchemy
dotenv
textract
faiss-cpu
argparse
logging
numpy
